test(user): #69 Test fixture를 각 domain 별로 분리
