+++ user/src/test/kotlin/com/damaba/user/adapter/inbound/auth/AuthControllerTest.kt
@@ -4,9 +4,9 @@ import com.damaba.user.adapter.inbound.auth.dto.OAuthLoginRequest
 import com.damaba.user.application.port.inbound.auth.OAuthLoginUseCase
 import com.damaba.user.config.ControllerTestConfig
 import com.damaba.user.domain.user.constant.LoginType
+import com.damaba.user.util.AuthFixture.createAuthToken
 import com.damaba.user.util.RandomTestUtils.Companion.randomString
-import com.damaba.user.util.TestFixture.createAuthToken
-import com.damaba.user.util.TestFixture.createUser
+import com.damaba.user.util.UserFixture.createUser
 import com.fasterxml.jackson.databind.ObjectMapper
 import io.mockk.every
 import io.mockk.mockk
