+++ user/src/test/kotlin/com/damaba/user/application/service/auth/OAuthLoginServiceTest.kt
@@ -11,9 +11,9 @@ import com.damaba.user.domain.auth.RefreshToken
 import com.damaba.user.domain.user.User
 import com.damaba.user.domain.user.constant.LoginType
 import com.damaba.user.property.AuthProperties
+import com.damaba.user.util.AuthFixture.createAuthToken
 import com.damaba.user.util.RandomTestUtils.Companion.randomString
-import com.damaba.user.util.TestFixture.createAuthToken
-import com.damaba.user.util.TestFixture.createUser
+import com.damaba.user.util.UserFixture.createUser
 import io.mockk.Runs
 import io.mockk.confirmVerified
 import io.mockk.every
