+++ user/src/test/kotlin/com/damaba/user/application/service/user/UserServiceTest.kt
@@ -17,7 +17,7 @@ import com.damaba.user.util.RandomTestUtils.Companion.randomBoolean
 import com.damaba.user.util.RandomTestUtils.Companion.randomLong
 import com.damaba.user.util.RandomTestUtils.Companion.randomString
 import com.damaba.user.util.RandomTestUtils.Companion.randomUrl
-import com.damaba.user.util.TestFixture.createUser
+import com.damaba.user.util.UserFixture.createUser
 import io.mockk.Runs
 import io.mockk.confirmVerified
 import io.mockk.every
